pyproject.toml
[tool.poetry]
name = "drf-api-permissions-postgres"
version = "0.1.0"
description = ""
authors = ["shahd1995913 <shahdalkhatib95@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
docker = "^5.0.3"
docker-compose = "^1.29.2"
sqlparse = "^0.4.2"
pytz = "^2021.3"
djangorestframework = "^3.12.4"
asgiref = "^3.4.1"
psycopg2-binary = "^2.9.2"
yml = "^0.0.1"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
flack = "^1.3.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
